Sketching out the framework of the CAA indexer
The idea is the main database has a queue that holds events about releases that
have cover art. If a release that has cover art is in some way updated, such as
the name changing, an event is produced. This project is a consumer of that
queue, and runs as a daemon. When it receieves events, it looks up the
information necessary to produce an index file, and sticks it on the Internet
Archive's backend servers.

Currently it produces part of the example.json file (release metadata).