Fix JSON utf8 encoding
With JSON::Any, the utf8 flag can have an inversed meaning depending on
what backend is used. With the modules in the current cpanfile.snapshot,
this is a problem, and encoding is broken in production. Use JSON::XS
directly to avoid this nonsense.

Some unrelated modules were also updated while running `carton install`
again.