Create a new DB handle for each index event
Whereas the previous code would halt the indexer with no hope of
recovery whenever there was a connection loss to the DB, this avoids
that issue by creating a new connection for each index event. This is
fine given the current speed of indexing and our use of a connection
pooler.